+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-11-17
+Authored by @bborysenko (_thank you, Borys!_)
+
+### Added
+- **Date-based search**: 
+  - supports filtering by note's creation and modification date
+  - supports relative dates, e.g., 'yesterday' or 'last week', etc or exact dates
+  - notes can be searched by date alone, without requiring a search term or tag
+
+
 ## [1.2.0] - 2025-11-02
 
 ### Added